+++ packages/types/src/index.ts
@@ -4,4 +4,4 @@ export * from './servers';
 export * from './indexed-db';
 export * from './utils';
 export * from './validation';
-export * from './hi-lo';
+export * from './lo-hi';
